# documentation/docs/content/openAPI/spring_backend-openapi.yaml
openapi: "3.0.3"
info:
  title: "spring_backend API"
  description: "spring_backend API"
  version: "1.0.0"
servers:
  - url: "https://spring_backend"
paths:
  /api/projects/{id}:
    put:
      summary: "PUT api/projects/{id}"
      operationId: "updateProject"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectDTO"
    delete:
      summary: "DELETE api/projects/{id}"
      operationId: "deleteProject"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectDTO"
    get:
      summary: "GET api/projects/{id}"
      operationId: "getProjectById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectDTO"
  /api/projects:
    post:
      summary: "POST api/projects"
      operationId: "createNewProject"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectDTO"
    get:
      summary: "GET api/projects"
      operationId: "getAllProjects"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/ProjectDTO"
  /api/projects/upload:
    post:
      summary: "POST api/projects/upload"
      operationId: "uploadProject"
      parameters:
        - name: "file"
          in: "query"
          required: true
          schema:
            type: "string"
            format: "binary"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProjectDTO"
  /api/projects/{id}/download:
    get:
      summary: "GET api/projects/{id}/download"
      operationId: "downloadProject"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Resource"
...
  /api/strategies/trend/{type}:
    get:
      summary: "GET api/strategies/trend/{type}"
      operationId: "getTrend"
      parameters:
        - name: "type"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsonNode"
...
  /api/strategies/season/{type}:
    get:
      summary: "GET api/strategies/season/{type}"
      operationId: "getSeason"
      parameters:
        - name: "type"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsonNode"
...
  /api/strategies/residual/{type}:
    get:
      summary: "GET api/strategies/residual/{type}"
      operationId: "getResidual"
      parameters:
        - name: "type"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsonNode"
...
  /api/projects/{projectId}/tracks/{trackId}:
    put:
      summary: "PUT api/projects/{projectId}/tracks/{trackId}"
      operationId: "updateTrack"
      parameters:
        - name: "projectId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrackDTO"
    delete:
      summary: "DELETE api/projects/{projectId}/tracks/{trackId}"
      operationId: "deleteProject"
      parameters:
        - name: "projectId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrackDTO"
    get:
      summary: "GET api/projects/{projectId}/tracks/{trackId}"
      operationId: "getTrackById"
      parameters:
        - name: "projectId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "trackId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrackDTO"
...
components:
  schemas:
    ProjectDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        sending:
          type: "boolean"
    Resource:
      type: "object"
      properties: { }
...
    JsonNode:
      type: "object"
      properties: { }
    DataTypeDTO:
      type: "object"
      properties:
        name:
          type: "string"
        description:
          type: "string"
        data_type:
          type: "string"
          enum:
            - "CHAR"
            - "FLOAT"
            - "ML"
            - "TSA"
            - "INTEGER"
            - "SLEEP"
        schema:
          type: "string"
        ui_schema:
          type: "string"
...
    TrackDTO:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        repeating:
          type: "boolean"
        name:
          type: "string"
        unit:
          type: "string"
        allowedDataTypes:
          type: "array"
          items:
            $ref: "#/components/schemas/DataTypeDTO"
...